+++ internal/alignfor/alignfor.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/mum4k/termdash/align"
 	"github.com/mum4k/termdash/internal/runewidth"
+	"github.com/mum4k/termdash/internal/wrap"
 )
 
 // hAlign aligns the given area in the rectangle horizontally.
@@ -89,10 +90,17 @@ func Rectangle(rect image.Rectangle, ar image.Rectangle, h align.Horizontal, v a
 // Text aligns the text within the given rectangle, returns the start point for the text.
 // For the purposes of the alignment this assumes that text will be trimmed if
 // it overruns the rectangle.
-// This only supports a single line of text, the text must not contain newlines.
+// This only supports a single line of text, the text must not contain non-printable characters,
+// allows empty text.
 func Text(rect image.Rectangle, text string, h align.Horizontal, v align.Vertical) (image.Point, error) {
 	if strings.ContainsRune(text, '\n') {
 		return image.ZP, fmt.Errorf("the provided text contains a newline character: %q", text)
+	}
+
+	if text != "" {
+		if err := wrap.ValidText(text); err != nil {
+			return image.ZP, fmt.Errorf("the provided text contains non printable character(s): %s", err)
+		}
 	}
 
 	cells := runewidth.StringWidth(text)